% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcenso_app.R
\name{arcenso_app}
\alias{arcenso_app}
\title{Launches a basic Shiny application to interactively browse and consult
Argentina's historical census data.}
\usage{
arcenso_app()
}
\value{
A Shiny app object.
}
\description{
Launches a basic Shiny application to interactively browse and consult
Argentina's historical census data.
}
\examples{
if (interactive()) {
  arcenso_app()
}

}
