% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcenso_gui.R
\name{arcenso_gui}
\alias{arcenso_gui}
\title{ARcenso Interactive Interface}
\usage{
arcenso_gui()
}
\value{
A Shiny app object.
}
\description{
Launches a basic Shiny application to interactively browse and consult
Argentina's historical census data.
}
\examples{
\dontrun{
arcenso_gui()
}
}
