% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repository.R
\name{check_repository}
\alias{check_repository}
\title{check_repository}
\usage{
check_repository(year = 1970, topic = NULL, geolvl = NULL)
}
\arguments{
\item{year}{census year for which you want to view the currently available tabulations}

\item{geolvl}{}
}
\description{
check_repository
}
