% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info_cuadros_arcenso}
\alias{info_cuadros_arcenso}
\title{Metadatos de los cuadros del Censo de Argentina}
\format{
Un data frame con 48 filas y 5 columnas:
\describe{
\item{Jurisdiccion}{Nombre de la provincia o jurisdicción (ASCII).}
\item{tema}{Categoría temática del censo.}
\item{Titulo}{Título descriptivo del cuadro estadístico.}
\item{Archivo}{Nombre del archivo asociado para descarga.}
\item{Anio}{Año del censo correspondiente.}
}
}
\source{
\url{https://www.indec.gob.ar/}
}
\usage{
info_cuadros_arcenso
}
\description{
Un dataset que contiene la relación entre jurisdicciones, temas y archivos
para facilitar la descarga y visualización interactiva.
}
\keyword{datasets}
