% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcenso_app.R
\name{arcenso_app}
\alias{arcenso_app}
\title{Launch ARcenso Shiny application}
\usage{
arcenso_app()
}
\value{
Launches the Shiny application (invisibly returns the result of
\code{\link[shiny:runApp]{shiny::runApp()}}).
}
\description{
Launches a Shiny application to interactively browse and consult
Argentina's historical census data.
}
\examples{
if (interactive()) {
  arcenso_app()
}

}
