% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repository.R
\name{check_repository}
\alias{check_repository}
\title{Check available census tables}
\usage{
check_repository(year = 1970, topic = NULL, geolvl = NULL)
}
\arguments{
\item{year}{Census year (1970 or 1980). Default is 1970.}

\item{topic}{Census topic of the tables (e.g., "EDUCACION").
If NULL, all available topics for the selected year will be shown.}

\item{geolvl}{Geographic disaggregation level (e.g., "Total del país").
If NULL, all jurisdictions will be included.}
}
\value{
A data frame containing the table titles and the internal names
required to access them via the get_census function.
}
\description{
Reports the census tables currently available in the package for consultation
based on the provided filters.
}
\examples{
check_repository(year = 1970, topic = "EDUCACION", geolvl = "Total del país")
}
