% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repository.R
\name{check_repository}
\alias{check_repository}
\title{check_repository
reports the tables currently available in the package}
\usage{
check_repository(year = 1970, topic = NULL, geolvl = NULL)
}
\arguments{
\item{year}{census year for which the user wants to view the currently available tabulations. The default is “1970”.}

\item{topic}{census topic of the tables Setting the parameter to NULL will download all available tables.}

\item{geolvl}{geographic disaggregation level, use “Total del país” for the overall results. Setting the parameter to NULL will download all available tables.}
}
\value{
Returns a data frame with the title of the tables that can be acces with the get_census function and the name given by that function to each table
}
\description{
check_repository
reports the tables currently available in the package
}
\examples{
check_repository( year = 1970, topic = "EDUCACION", geolvl = "Total del país")
}
