% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{Get census data tables}
\usage{
get_census(year = 1970, topic = NULL, geolvl = NULL)
}
\arguments{
\item{year}{Census year (1970 or 1980). Default is 1970.}

\item{topic}{Census topic of the tables (e.g., "EDUCACION").
If NULL, all available topics for the year will be retrieved.}

\item{geolvl}{Geographic disaggregation level (e.g., "Total del país").
If NULL, all jurisdictions will be retrieved.}
}
\value{
A list of data frames containing the requested census tables.
}
\description{
Retrieves a list of curated data tables from Argentina's official population
censuses based on the specified filters.
}
